+++ Sea Battle/Sea Battle/ChangeColAndPerevirkaCoordinate.h
@@ -1,6 +1,6 @@
 #pragma once
 #define SIZEBOARD 10
 
-int ChangeColPlayer(char ColPlayer);
+short int ChangeColPlayer(char ColPlayer);
 
-bool PerevirkaNaCoordinate(char Board[SIZEBOARD][SIZEBOARD], int RowPlayer, int ColPlayer);
+bool PerevirkaNaCoordinate(char Board[SIZEBOARD][SIZEBOARD],short int RowPlayer,short int ColPlayer);
