+++ Sea Battle/Sea Battle/HitOrMissAndFarb.h
@@ -1,9 +1,9 @@
 #define SIZEBOARD 10
 
-bool HitOrMiss(char Board[SIZEBOARD][SIZEBOARD], int RowPlayer, int ColPlayer);
+bool HitOrMiss(char Board[SIZEBOARD][SIZEBOARD],short int RowPlayer,short int ColPlayer);
 
-bool IsSunkenShip(char BoardPlayer[SIZEBOARD][SIZEBOARD], char BoardShip[SIZEBOARD][SIZEBOARD], int RowPlayer, int ColPlayer);
+bool IsSunkenShip(char BoardPlayer[SIZEBOARD][SIZEBOARD], char BoardShip[SIZEBOARD][SIZEBOARD],short int RowPlayer,short int ColPlayer);
 
-void FarbSunkenShip(char BoardPlayer[SIZEBOARD][SIZEBOARD], char BoardShip[SIZEBOARD][SIZEBOARD],int RowPlayer, int ColPlayer);
+void FarbSunkenShip(char BoardPlayer[SIZEBOARD][SIZEBOARD], char BoardShip[SIZEBOARD][SIZEBOARD],short int RowPlayer,short int ColPlayer);
 
-void ChangeSymbolInCell(char BoardPlayer[SIZEBOARD][SIZEBOARD], char BoardShip[SIZEBOARD][SIZEBOARD], int RowPlayer, int ColPlayer);
+void ChangeSymbolInCell(char BoardPlayer[SIZEBOARD][SIZEBOARD], char BoardShip[SIZEBOARD][SIZEBOARD], short int RowPlayer, short int ColPlayer);
