+++ Sea Battle/Sea Battle/WinnerPlayer.h
@@ -1,6 +1,6 @@
 #pragma once
 #define SIZEBOARD 10
 
-short int CountSunkenCell(char BoardPlayer[SIZEBOARD][SIZEBOARD], int RowPlayer, int ColPlayer);
+short int CountSunkenCell(char BoardPlayer[SIZEBOARD][SIZEBOARD],short int RowPlayer,short int ColPlayer);
 
-bool WinnerPlayer(int CountSunkenCell);
+bool WinnerPlayer(short int CountSunkenCell);
